power: add static_casts for printf
The new version of Clang warns when a value of `enum class` type is
passed to a `printf`-like function without an explicit cast. This adds
the casts that Clang suggests.

BUG=b:305714217
TEST=FEATURES=test emerge-hana

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5013236
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>

+++ power_manager/powerd/policy/state_controller_test.cc
@@ -56,12 +56,12 @@ const char kNoActions[] = "";
 
 // Return a string representation of DimEvent.
 std::string GetDimEventString(metrics::DimEvent sample) {
-  return base::StringPrintf("DimEvent-%d", sample);
+  return base::StringPrintf("DimEvent-%d", static_cast<int>(sample));
 }
 
 // Return a string representation of LockEvent.
 std::string GetLockEventString(metrics::LockEvent sample) {
-  return base::StringPrintf("LockEvent-%d", sample);
+  return base::StringPrintf("LockEvent-%d", static_cast<int>(sample));
 }
 
 // Return a string representation of event name and duration.
